+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/Scheduling/IPeriodicRunningJob.cs
@@ -1,0 +1,8 @@
+﻿namespace ImoutoRebirth.Room.UI.Scheduling;
+
+public interface IPeriodicRunningJob
+{
+    TimeSpan PeriodDelay { get; }
+    
+    Task Run(CancellationToken token);
+}