+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
-﻿using ImoutoRebirth.Common.Quartz.Extensions;
-using ImoutoRebirth.Room.UI.Quartz;
+﻿using ImoutoRebirth.Room.UI.Scheduling;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace ImoutoRebirth.Room.UI;
@@ -8,7 +7,10 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddRoomUi(this IServiceCollection services)
     {
-        services.AddQuartzJob<OverseeJob, OverseeJob.Description>();
+        services.Configure<PeriodicRunnerOptions>(x => x.AddJob<OverseeJob>());
+        services.AddTransient<OverseeJob>();
+        services.AddHostedService<PeriodicRunnerHostedService>();
+        
         services.AddMemoryCache();
 
         return services;